Fix building with qreal float

+++ src/bodymovin/beziereasing.cpp
@@ -39,7 +39,7 @@ void BezierEasing::addCubicBezierSegment(const QPointF &c1, const QPointF &c2, c
 qreal BezierEasing::valueForProgress(qreal progress) const
 {
     qreal res = mBezier.pointAt(tForX(progress)).y();
-    return qBound(0.0, res, 1.0);
+    return qBound(qreal(0.0), res, qreal(1.0));
 }
 
 qreal BezierEasing::tForX(qreal x) const
